windows golangci-lint fixes for 'agent' cmd
