+++ cmd/agent/subcommands/integrations/integrations_darwin.go
@@ -11,26 +11,7 @@ package integrations
 import (
 	"fmt"
 	"os"
-	"path/filepath"
 )
-
-const (
-	pythonBin = "python"
-)
-
-func getRelPyPath(cliParams *cliParams) string {
-	return filepath.Join("embedded", "bin", fmt.Sprintf("%s%s", pythonBin, cliParams.pythonMajorVersion))
-}
-
-func getRelChecksPath(cliParams *cliParams) (string, error) {
-	err := detectPythonMinorVersion(cliParams)
-	if err != nil {
-		return "", err
-	}
-
-	pythonDir := fmt.Sprintf("%s%s.%s", pythonBin, cliParams.pythonMajorVersion, pythonMinorVersion)
-	return filepath.Join("embedded", "lib", pythonDir, "site-packages", "datadog_checks"), nil
-}
 
 func validateUser(allowRoot bool) error {
 	if os.Geteuid() != 0 {
