+++ cmd/system-probe/config/config_windows.go
@@ -11,7 +11,6 @@ package config
 import (
 	"fmt"
 	"net"
-	"path/filepath"
 
 	"github.com/DataDog/datadog-agent/pkg/util/winutil"
 )
@@ -25,15 +24,13 @@ const (
 )
 
 var (
-	defaultConfigDir              = "c:\\programdata\\datadog\\"
-	defaultSystemProbeLogFilePath = "c:\\programdata\\datadog\\logs\\system-probe.log"
+	defaultConfigDir = "c:\\programdata\\datadog\\"
 )
 
 func init() {
 	pd, err := winutil.GetProgramDataDir()
 	if err == nil {
 		defaultConfigDir = pd
-		defaultSystemProbeLogFilePath = filepath.Join(pd, "logs", "system-probe.log")
 	}
 }
