+++ cmd/systray/doconfigure.go
@@ -61,7 +61,7 @@ func doConfigure() error {
 	csrfToken, err := util.DoGet(c, urlstr, util.LeaveConnectionOpen)
 	if err != nil {
 		var errMap = make(map[string]string)
-		json.Unmarshal(csrfToken, &errMap)
+		err = json.Unmarshal(csrfToken, &errMap)
 		if e, found := errMap["error"]; found {
 			err = fmt.Errorf(e)
 		}
